refactor(epost-tracker): contract uplift + REST→direct calls
- Phase 35: no web/models/ to lift (zero Swagger-generated DTOs)
- Manual contract lift: 20 types to dristi-common/contract/epost/
  (ChannelResponse/Message/Name/Report, EPost*/DeliveryStatus/ExcelSheetType,
  Pagination/Sort/Order, Address/Coordinate/AdditionalFields/Field)
- @NamedInterface("contract-epost") added for Spring Modulith boundary
- @ApplicationModule marker added (epost/package-info.java)
- REST calls deferred: SummonsHost (summons-svc pending), PdfServiceHost (platform)
