+++ dristi-monolith/domain-integration/src/main/java/org/pucar/dristi/integration/epost/internal/controller/EPostController.java
@@ -7,6 +7,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
 import org.egov.common.contract.response.ResponseInfo;
+import org.pucar.dristi.common.contract.epost.*;
 import org.pucar.dristi.integration.epost.internal.model.*;
 import org.pucar.dristi.integration.epost.internal.service.EPostService;
 import org.pucar.dristi.integration.epost.internal.service.ExcelService;
