--- dristi-monolith/domain-integration/src/main/java/org/pucar/dristi/integration/epost/internal/model/ChannelName.java
@@ -1,8 +1,0 @@
-package org.pucar.dristi.integration.epost.internal.model;
-
-public enum ChannelName {
-
-    POLICE,
-
-    Post
-}
