--- dristi-monolith/domain-integration/src/main/java/org/pucar/dristi/integration/epost/internal/model/Order.java
@@ -1,30 +1,0 @@
-package org.pucar.dristi.integration.epost.internal.model;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonValue;
-
-public enum Order {
-	ASC("asc"), DESC("desc");
-
-	private String value;
-
-	Order(String value) {
-		this.value = value;
-	}
-
-	@Override
-	@JsonValue
-	public String toString() {
-		return String.valueOf(value);
-	}
-
-	@JsonCreator
-	public static Order fromValue(String text) {
-		for (Order b : Order.values()) {
-			if (String.valueOf(b.value).equals(text)) {
-				return b;
-			}
-		}
-		return null;
-	}
-}
