+++ dristi-monolith/domain-integration/src/main/java/org/pucar/dristi/integration/epost/internal/model/RespondentDetails.java
@@ -6,6 +6,7 @@ import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 import lombok.Data;
+import org.pucar.dristi.common.contract.epost.Address;
 import org.springframework.validation.annotation.Validated;
 
 @Validated
