+++ dristi-monolith/domain-integration/src/main/java/org/pucar/dristi/integration/epost/internal/model/UpdateSummonsRequest.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.egov.common.contract.request.RequestInfo;
+import org.pucar.dristi.common.contract.epost.ChannelReport;
 
 @Data
 @Builder
