+++ dristi-monolith/domain-integration/src/main/java/org/pucar/dristi/integration/epost/internal/model/WitnessDetails.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.pucar.dristi.common.contract.epost.Address;
 import org.springframework.validation.annotation.Validated;
 
 @Validated
