+++ dristi-monolith/domain-integration/src/main/java/org/pucar/dristi/integration/epost/internal/repository/EPostQueryBuilder.java
@@ -1,8 +1,8 @@
 package org.pucar.dristi.integration.epost.internal.repository;
 
 import lombok.extern.slf4j.Slf4j;
-import org.pucar.dristi.integration.epost.internal.model.EPostTrackerSearchCriteria;
-import org.pucar.dristi.integration.epost.internal.model.Pagination;
+import org.pucar.dristi.common.contract.epost.EPostTrackerSearchCriteria;
+import org.pucar.dristi.common.contract.epost.Pagination;
 import org.springframework.stereotype.Component;
 import org.springframework.util.CollectionUtils;
 import org.springframework.util.ObjectUtils;
