+++ dristi-monolith/domain-integration/src/main/java/org/pucar/dristi/integration/epost/internal/repository/EPostRepository.java
@@ -1,10 +1,10 @@
 package org.pucar.dristi.integration.epost.internal.repository;
 
 import lombok.extern.slf4j.Slf4j;
-import org.pucar.dristi.integration.epost.internal.model.EPostResponse;
-import org.pucar.dristi.integration.epost.internal.model.EPostTracker;
-import org.pucar.dristi.integration.epost.internal.model.EPostTrackerSearchCriteria;
-import org.pucar.dristi.integration.epost.internal.model.Pagination;
+import org.pucar.dristi.common.contract.epost.EPostResponse;
+import org.pucar.dristi.common.contract.epost.EPostTracker;
+import org.pucar.dristi.common.contract.epost.EPostTrackerSearchCriteria;
+import org.pucar.dristi.common.contract.epost.Pagination;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
