+++ dristi-monolith/domain-integration/src/main/java/org/pucar/dristi/integration/epost/internal/repository/EPostRowMapper.java
@@ -4,10 +4,10 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.egov.common.contract.models.AuditDetails;
-import org.pucar.dristi.integration.epost.internal.model.AdditionalFields;
-import org.pucar.dristi.integration.epost.internal.model.Address;
-import org.pucar.dristi.integration.epost.internal.model.DeliveryStatus;
-import org.pucar.dristi.integration.epost.internal.model.EPostTracker;
+import org.pucar.dristi.common.contract.epost.AdditionalFields;
+import org.pucar.dristi.common.contract.epost.Address;
+import org.pucar.dristi.common.contract.epost.DeliveryStatus;
+import org.pucar.dristi.common.contract.epost.EPostTracker;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
