+++ dristi-monolith/domain-integration/src/main/java/org/pucar/dristi/integration/epost/internal/util/SummonsUtil.java
@@ -5,9 +5,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.egov.tracer.model.CustomException;
 import org.pucar.dristi.integration.epost.internal.config.EPostConfiguration;
-import org.pucar.dristi.integration.epost.internal.model.ChannelReport;
-import org.pucar.dristi.integration.epost.internal.model.DeliveryStatus;
-import org.pucar.dristi.integration.epost.internal.model.EPostRequest;
+import org.pucar.dristi.common.contract.epost.ChannelReport;
+import org.pucar.dristi.common.contract.epost.DeliveryStatus;
+import org.pucar.dristi.common.contract.epost.EPostRequest;
 import org.pucar.dristi.integration.epost.internal.model.UpdateSummonsRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpEntity;
