+++ dristi-monolith/domain-integration/src/main/java/org/pucar/dristi/integration/epost/internal/validator/EPostUserValidator.java
@@ -3,7 +3,7 @@ package org.pucar.dristi.integration.epost.internal.validator;
 import lombok.extern.slf4j.Slf4j;
 import org.egov.common.contract.request.RequestInfo;
 import org.pucar.dristi.integration.epost.internal.config.MdmsDataConfig;
-import org.pucar.dristi.integration.epost.internal.model.EPostTrackerSearchRequest;
+import org.pucar.dristi.common.contract.epost.EPostTrackerSearchRequest;
 import org.pucar.dristi.common.util.MdmsUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
