+++ dristi-monolith/domain-integration/src/main/java/org/pucar/dristi/integration/epost/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * Epost subdomain — India Post e-post tracking integration.
+ *
+ * <p>Marked as a Spring Modulith application module. No cross-subdomain
+ * callers exist today; this marker is here for consistency so that when
+ * a caller appears, the boundary is already enforced.
+ */
+@org.springframework.modulith.ApplicationModule(displayName = "Epost")
+package org.pucar.dristi.integration.epost;
