+++ dristi-monolith/domain-integration/src/test/java/org/pucar/dristi/integration/epost/internal/repository/EPostRowMapperTest.java
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.egov.common.contract.models.AuditDetails;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.pucar.dristi.integration.epost.internal.model.AdditionalFields;
-import org.pucar.dristi.integration.epost.internal.model.DeliveryStatus;
-import org.pucar.dristi.integration.epost.internal.model.EPostTracker;
+import org.pucar.dristi.common.contract.epost.AdditionalFields;
+import org.pucar.dristi.common.contract.epost.DeliveryStatus;
+import org.pucar.dristi.common.contract.epost.EPostTracker;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
