+++ dristi-monolith/domain-integration/src/test/java/org/pucar/dristi/integration/epost/internal/service/EPostServiceTest.java
@@ -10,6 +10,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.pucar.dristi.integration.epost.internal.config.EPostConfiguration;
 import org.pucar.dristi.common.kafka.Producer;
+import org.pucar.dristi.common.contract.epost.*;
 import org.pucar.dristi.integration.epost.internal.model.*;
 import org.pucar.dristi.integration.epost.internal.repository.EPostRepository;
 import org.pucar.dristi.integration.epost.internal.util.EpostUtil;
