+++ dristi-monolith/dristi-common/src/main/java/org/pucar/dristi/common/contract/epost/ChannelMessage.java
@@ -1,9 +1,8 @@
-package org.pucar.dristi.integration.epost.internal.model;
+package org.pucar.dristi.common.contract.epost;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 import lombok.Data;
-
 
 @Data
 @Builder
