+++ dristi-monolith/dristi-common/src/main/java/org/pucar/dristi/common/contract/epost/ChannelName.java
@@ -1,0 +1,7 @@
+package org.pucar.dristi.common.contract.epost;
+
+public enum ChannelName {
+
+    POLICE,
+    Post
+}
