+++ dristi-monolith/dristi-common/src/main/java/org/pucar/dristi/common/contract/epost/ChannelReport.java
@@ -1,4 +1,4 @@
-package org.pucar.dristi.integration.epost.internal.model;
+package org.pucar.dristi.common.contract.epost;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.*;
@@ -24,5 +24,4 @@ public class ChannelReport {
 
     @JsonProperty("additionalFields")
     private AdditionalFields additionalFields;
-
 }
