+++ dristi-monolith/dristi-common/src/main/java/org/pucar/dristi/common/contract/epost/ChannelResponse.java
@@ -1,4 +1,4 @@
-package org.pucar.dristi.integration.epost.internal.model;
+package org.pucar.dristi.common.contract.epost;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.Valid;
@@ -8,23 +8,17 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.egov.common.contract.response.ResponseInfo;
 
-import java.util.List;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class EPostResponse {
+public class ChannelResponse {
 
     @JsonProperty("ResponseInfo")
     @Valid
     private ResponseInfo responseInfo = null;
 
-    @JsonProperty("EPostTracker")
+    @JsonProperty("channelMessage")
     @Valid
-    private List<EPostTracker> ePostTrackers = null;
-
-    @JsonProperty("pagination")
-    @Valid
-    private Pagination pagination = null;
+    private ChannelMessage channelMessage = null;
 }
