+++ dristi-monolith/dristi-common/src/main/java/org/pucar/dristi/common/contract/epost/EPostRequest.java
@@ -1,4 +1,4 @@
-package org.pucar.dristi.integration.epost.internal.model;
+package org.pucar.dristi.common.contract.epost;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.Valid;
