+++ dristi-monolith/dristi-common/src/main/java/org/pucar/dristi/common/contract/epost/EPostResponse.java
@@ -1,4 +1,4 @@
-package org.pucar.dristi.integration.epost.internal.model;
+package org.pucar.dristi.common.contract.epost;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.Valid;
@@ -8,17 +8,23 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.egov.common.contract.response.ResponseInfo;
 
+import java.util.List;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class ChannelResponse {
+public class EPostResponse {
 
     @JsonProperty("ResponseInfo")
     @Valid
     private ResponseInfo responseInfo = null;
 
-    @JsonProperty("channelMessage")
+    @JsonProperty("EPostTracker")
     @Valid
-    private ChannelMessage channelMessage = null;
+    private List<EPostTracker> ePostTrackers = null;
+
+    @JsonProperty("pagination")
+    @Valid
+    private Pagination pagination = null;
 }
