+++ dristi-monolith/dristi-common/src/main/java/org/pucar/dristi/common/contract/epost/ExcelSheetType.java
@@ -1,11 +1,8 @@
-package org.pucar.dristi.integration.epost.internal.model;
+package org.pucar.dristi.common.contract.epost;
 
 public enum ExcelSheetType {
 
     PENDING_BOOKING_TAB,
-
     REPORTS_TAB,
-
     MONTHLY_REPORT_EMAIL
-
 }
