+++ dristi-monolith/dristi-common/src/main/java/org/pucar/dristi/common/contract/epost/Pagination.java
@@ -1,4 +1,4 @@
-package org.pucar.dristi.integration.epost.internal.model;
+package org.pucar.dristi.common.contract.epost;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -8,9 +8,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.validation.annotation.Validated;
 
-/**
- * Pagination details
- */
 @Schema(description = "Pagination details")
 @Validated
 @Data
@@ -18,12 +15,13 @@ import org.springframework.validation.annotation.Validated;
 @NoArgsConstructor
 @Builder
 public class Pagination {
-	@JsonProperty("totalCount")
-	private Integer totalCount = null;
 
-	@JsonProperty("sortBy")
-	private Sort sortBy = null;
+    @JsonProperty("totalCount")
+    private Integer totalCount = null;
 
-	@JsonProperty("orderBy")
-	private Order orderBy = null;
+    @JsonProperty("sortBy")
+    private Sort sortBy = null;
+
+    @JsonProperty("orderBy")
+    private Order orderBy = null;
 }
